requirements.txt
tensorflow~=2.18.0
numpy~=2.0.2
pandas~=2.2.3
scikit-learn~=1.6.0
keras~=3.7.0
pillow~=11.0.0